requirements.txt
streamlit
pandas
pyyaml
plotly
requests
ccxt
